finished 4.21

+++ blog_list/server/controllers/blogs.js
@@ -7,13 +7,14 @@ const jwt = require('jsonwebtoken')
 const User = require('../models/user')
 const Blog = require('../models/blog')
 
-const getTokenForm = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+// This function has moved to middleware
+// const getTokenForm = request => {
+//   const authorization = request.get('authorization')
+//   if (authorization && authorization.startsWith('Bearer ')) {
+//     return authorization.replace('Bearer ', '')
+//   }
+//   return null
+// }
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -34,7 +35,8 @@ blogsRouter.get('/:id', async (request, response, next) => {
 // Posting a blog
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body 
-  const decodedToken = jwt.verify(getTokenForm(request), process.env.SECRET)
+  // const decodedToken = jwt.verify(getTokenForm(request), process.env.SECRET)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
@@ -73,8 +75,25 @@ blogsRouter.put('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog) {
+    if (blog.user.toString() === user._id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(400).json({ error: 'blog does not belong to user' })
+    }
+  } else {
+    response.status(404).end()
+  }
 })
 
 module.exports = blogsRouter
